factor out mention hud logic into a HOC
